MAINT(tests): Consolidate all test specs in a single /test folder
