+++ test/api/users.spec.js
@@ -2,8 +2,8 @@
 
 const {expect} = require('chai')
 const request = require('supertest')
-const db = require('../db')
-const app = require('../index')
+const db = require('../../server/db')
+const app = require('../../server/index')
 const User = db.model('user')
 
 describe('User routes', () => {
